backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.11.5
python-multipart==0.0.12
python-dotenv==1.1.0

# AI features
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
googleapis-common-protos==1.72.0
